move before pre gyp
