+++ scripts/postinstall.js
@@ -112,13 +112,3 @@ if (require.main === module) {
     }
   })
 }
-
-// https://github.com/nodejs/nan/issues/978
-const filePath = path.join(__dirname, '../node_modules/nan/nan.h')
-const fileContent = fs.readFileSync(filePath, 'utf8')
-const updatedContent = fileContent.replace(
-  /#include "nan_scriptorigin.h"/,
-  '// #include "nan_scriptorigin.h"',
-)
-
-fs.writeFileSync(filePath, updatedContent, 'utf8')
